% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotation}
\alias{rotation}
\title{Visuomotor rotation data.}
\format{A matrix with 80 rows (obs) and 40 columns (subjects):}
\usage{
rotation
}
\description{
A dataset containing initial reach angle data from 40 subjects.
The rotation occurs at observation #9.
}
\keyword{datasets}
