% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_mcmc}
\alias{simulate_mcmc}
\title{Simulate an MCMC-based model for visuomotor rotation.}
\usage{
simulate_mcmc(optimal, cost_width, forget_rate, proposal_variance)
}
\arguments{
\item{optimal}{A numeric vector containing the optimal reach direction.}

\item{cost_width}{Width of the cost function.}

\item{forget_rate}{Forgetting rate.}

\item{proposal_variance}{Width of the proposal distribution.}
}
\description{
Simulate an MCMC-based model for visuomotor rotation.
}
\examples{
optimal <- c(rep(0, 8), rep(30, 80))
sims <- simulate_mcmc(optimal = optimal, 
                      cost_width = 5,
                      forget_rate = 0.9,
                      proposal_variance = 4.5)
plot(optimal, type = 'l')
lines(sims, col = 'red')
}
