% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_ssm}
\alias{simulate_ssm}
\title{Simulate a state space model for visuomotor rotation.}
\usage{
simulate_ssm(optimal, learn_rate, forget_rate, process_variance,
  execution_variance)
}
\arguments{
\item{optimal}{A numeric vector containing the optimal reach direction.}

\item{learn_rate}{Learning rate.}

\item{forget_rate}{Forgetting rate.}

\item{process_variance}{State noise.}

\item{execution_variance}{Execution noise.}
}
\description{
Simulate a state space model for visuomotor rotation.
}
\examples{
optimal <- c(rep(0, 8), rep(30, 80))
sims <- simulate_ssm(optimal = optimal, 
                     learn_rate = 0.15,
                     forget_rate = 0.94,
                     process_variance = 1,
                     execution_variance = 5)
plot(optimal, type = 'l')
lines(sims, col = 'red')
}
